[#119] fix : put 메소드로 파일업로드가 불가하여 post로 변경
230801 김수현

+++ BE/src/main/java/com/ssafy/withview/repository/dto/ChannelDto.java
@@ -18,7 +18,7 @@ public class ChannelDto {
 	private long seq;
 	String name;
 	int limitPeople;
-	int serverSeq;
+	long serverSeq;
 	private String backgroundImgSearchName;
 	private String backgroundImgOriginalName;
