+++ BE/src/main/java/com/ssafy/withview/repository/dto/ServerDto.java
@@ -19,7 +19,7 @@ public class ServerDto {
 	private long seq;
 	private String name;
 	private int limitChannel;
-	private int hostSeq;
+	private long hostSeq;
 	private String backgroundImgSearchName;
 	private String backgroundImgOriginalName;
