+++ BE/src/main/java/com/ssafy/withview/service/ChannelService.java
@@ -8,7 +8,7 @@ import org.springframework.web.multipart.MultipartFile;
 import java.util.List;
 
 public interface ChannelService {
-	public List<ChannelDto> findAllChannelsByServerSeq(int serverSeq) ;
+	public List<ChannelDto> findAllChannelsByServerSeq(long serverSeq) ;
 	public ChannelDto insertChannel(ChannelDto channelDto, MultipartFile multipartFile) throws Exception;
 	public ChannelDto updateChannel(ChannelDto channelDto, MultipartFile multipartFile) throws Exception;
 	public void deleteChannel(long channelSeq,long userSeq) throws Exception;
